refactor: move URL-aware filter validation to services layer
- Create services/filter_validation.py with validate_filters_for_url
- Remove service import from schemas/offer_filter.py (layer violation fix)
- Update API to import from services
- Schemas now contain only pure Pydantic models + basic structure validation
